requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-storage-blob
fitz
azure-cosmos
gremlinpython
azure-ai-formrecognizer
openai
azure-core
azure-cosmos
azure-search-documents
langchain-openai